Create new and move existing navigation test helper methods
