+++ tests/cypress/support/helpers.js
@@ -81,6 +81,257 @@ const convertDateToDisplayDate = (date) => {
   return `${dateString} between ${timeString}`
 }
 
+const navigateToCommunalPage = () => {
+  cy.visit('http://localhost:3000/report-repair/');
+  makeSelectionAndClickButton('Something else');
+}
+
+const navigateToPostcodePage = () => {
+  cy.visit('http://localhost:3000/report-repair/');
+  makeSelectionAndClickButton('Something else');
+  cy.get('[data-cy=communal]', {timeout: 10000}).then(($loadedSection) => {
+    makeSelectionAndClickButton('No');
+  });
+  cy.get('[data-cy=postcode]', {timeout: 10000})
+}
+
+const navigateToAddress = () => {
+  cy.visit('http://localhost:3000/');
+  cy.contains('Start now').click();
+  cy.contains('Something else').click();
+  cy.get('button').click();
+  cy.get('[data-cy=communal]', {timeout: 10000}).then(($loadedSection) => {
+    cy.contains('No').click();
+  });
+}
+
+const navigateToBestDescriptionPage = () => {
+  cy.visit('http://localhost:3000/report-repair/');
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'priority-list',
+    option:'Something else'
+  })
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'communal', option:'No'
+  })
+  navigateToPageTypeInputTextAndContinue({
+    page: 'postcode', inputText:'SW1A 2AA'
+  })
+  const address = '1 Downing Street, London, SW1A 2AA';
+  cy.get('[data-cy=address]', {timeout: 10000}).then(() => {
+    cy.get('select').select(address)
+    cy.get('button').click();
+  });
+  makeSelectionAndClickButton('Kitchen');
+  makeSelectionAndClickButton('Damp or mould');
+
+}
+
+const navigateToRepairAvailabilityPage = () => {
+  const address = '1 Downing Street, London, SW1A 2AA';
+  const repairDescription = 'Eius postea venit saepius arcessitus.'
+  const phoneNumber = '02085548333';
+  const email = 'harrypotter@hogwarts.com';
+
+  cy.visit('http://localhost:3000/report-repair/');
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'priority-list',
+    option: 'Something else'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'communal', option: 'No'
+  })
+
+  navigateToPageTypeInputTextAndContinue({
+    page: 'postcode', inputText: 'SW1A 2AA'
+  })
+
+  cy.get('[data-cy=address]', {timeout: 10000}).then(() => {
+    cy.get('select').select(address)
+    cy.get('button').click();
+  });
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-location', option: 'Kitchen'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-problem', option: 'Damaged worktop'
+  })
+
+  cy.get('[data-cy=repair-description]', {timeout: 10000}).then(() => {
+    cy.get('textarea').type(repairDescription);
+    cy.get('button').contains('Continue').click();
+  });
+
+  cy.get('[data-cy=repair-image-upload]', {timeout: 10000}).then(() => {
+    cy.get('button').contains('Continue').click();
+  });
+
+  navigateToPageTypeInputTextAndContinue({
+    page: 'contact-person',
+    inputText: phoneNumber
+  })
+
+  cy.get('[data-cy=contact-details]', {timeout: 10000}).then(() => {
+    cy.get('input#contactDetails-1').click().then(() => {
+      cy.get('input#contactDetails-email').type(email);
+    })
+    cy.get('button').click();
+  });
+}
+
+const navigateToSummaryPage = () => {
+  let timeSlot = ''
+  const address = '1 Downing Street, London, SW1A 2AA';
+  const repairDescription = 'Eius postea venit saepius arcessitus.'
+  const phoneNumber = '02085548333';
+  const email = 'harrypotter@hogwarts.com';
+
+  cy.visit('http://localhost:3000/report-repair/');
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'priority-list',
+    option: 'Something else'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'communal', option: 'No'
+  })
+
+  navigateToPageTypeInputTextAndContinue({
+    page: 'postcode', inputText: 'SW1A 2AA'
+  })
+
+  cy.get('[data-cy=address]', {timeout: 10000}).then(() => {
+    cy.get('select').select(address)
+    cy.get('button').click();
+  });
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-location', option: 'Kitchen'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-problem',
+    option: 'Cupboards, including damaged cupboard doors'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-problem-best-description', option: 'Hanging door'
+  })
+
+  cy.get('[data-cy=repair-description]', {timeout: 10000}).then(() => {
+    cy.get('textarea').type(repairDescription);
+    cy.get('button').contains('Continue').click();
+  });
+
+  cy.get('[data-cy=repair-image-upload]', {timeout: 10000}).then(() => {
+    cy.get('button').contains('Continue').click();
+  });
+
+  navigateToPageTypeInputTextAndContinue({
+    page: 'contact-person',
+    inputText: phoneNumber
+  })
+
+  cy.get('[data-cy=contact-details]', {timeout: 10000}).then(() => {
+    cy.get('input#contactDetails-1').click().then(() => {
+      cy.get('input#contactDetails-email').type(email);
+    })
+    cy.get('button').click();
+  });
+
+  cy.get('[data-cy=repair-availability]', {timeout: 10000}).then(() => {
+    cy.get('[data-cy=availability-slot-0-0]').invoke('val').then(value => {
+      timeSlot = value;
+    })
+    cy.get('[data-cy=availability-slot-0-0]').click();
+    cy.get('button').click();
+  });
+
+  return () => timeSlot;
+}
+
+const completeJourney = (useContactPhone = false, contactValue) => {
+  const address = '1 Downing Street, London, SW1A 2AA';
+  const repairDescription = 'Eius postea venit saepius arcessitus.'
+  const phoneNumber = '07512345678';
+
+  cy.visit('http://localhost:3000/report-repair/');
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'priority-list',
+    option: 'Something else'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'communal', option: 'No'
+  })
+
+  navigateToPageTypeInputTextAndContinue({
+    page: 'postcode', inputText: 'SW1A 2AA'
+  })
+
+  cy.get('[data-cy=address]', {timeout: 10000}).then(() => {
+    cy.get('select').select(address)
+    cy.get('button').click();
+  });
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-location', option: 'Kitchen'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-problem',
+    option: 'Cupboards, including damaged cupboard doors'
+  })
+
+  navigateToPageSelectRadioOptionAndContinue({
+    page: 'repair-problem-best-description', option: 'Hanging door'
+  })
+
+  cy.get('[data-cy=repair-description]', {timeout: 10000}).then(() => {
+    cy.get('textarea').type(repairDescription);
+    cy.get('button').contains('Continue').click();
+  });
+
+  cy.get('[data-cy=repair-image-upload]', {timeout: 10000}).then(() => {
+    cy.get('button').contains('Continue').click();
+  });
+
+  navigateToPageTypeInputTextAndContinue({
+    page: 'contact-person',
+    inputText: phoneNumber
+  })
+
+  cy.get('[data-cy=contact-details]', {timeout: 10000}).then(() => {
+    if (useContactPhone) {
+      cy.contains('Text message (recommended)').click().then(() => {
+        cy.get('input#contactDetails-text').type(contactValue);
+      })
+    } else {
+      cy.contains('Email').click().then(() => {
+        cy.get('input#contactDetails-email').type(contactValue);
+      })
+    }
+    cy.get('button').click();
+  });
+
+  cy.get('[data-cy=repair-availability]', {timeout: 10000}).then(() => {
+    cy.get('[data-cy=availability-slot-0-0]').click();
+    cy.get('button').click();
+  });
+
+  cy.get('[data-cy=summary]', {timeout: 10000}).then(() => {
+    cy.get('button').click();
+  });
+
+  return () => { }
+}
+
 const navigateToLocation = () => {
   intercept_address_search();
   intercept_repair_triage();
@@ -125,7 +376,14 @@ export {
   convertDateToDisplayDate,
   intercept_save_repair,
   continueOnPage,
+  navigateToCommunalPage,
+  navigateToPostcodePage,
+  navigateToAddress,
   navigateToLocation,
+  navigateToBestDescriptionPage,
+  navigateToRepairAvailabilityPage,
+  navigateToSummaryPage,
+  completeJourney,
   makeSelectionAndClickButton,
   checkIfSelectionGoesToCorrectUrl,
   intercept_repair_triage
